Move geotiff export functions to DEADataHandling.py
